Split `source` module into submodules
